Getting real deep in it.
There’s some pretty messy stuff in here, especially code copy pasta in
the controls module. But the html looks very fancy, I think. So very
organized in the web page. So far, it looks like the nicest way to
handle scaling is to just automatically do it behind the scenes when
getting the sizes from the Settings module.